# Optimize for size and performance the most common log message