--- a/include/zephyr/logging/log_msg.h
+++ b/include/zephyr/logging/log_msg.h
@@ -83,7 +83,6 @@ struct log_msg_hdr {
 	void *tid;
 #endif
 };
-
 /* Messages are aligned to alignment required by cbprintf package. */
 #define Z_LOG_MSG_ALIGNMENT CBPRINTF_PACKAGE_ALIGNMENT
 
@@ -141,6 +140,9 @@ enum z_log_msg_mode {
 	 * more code size.
 	 */
 	Z_LOG_MSG_MODE_ZERO_COPY,
+
+	/* Mode optimized for simple messages with 0 to 2 32 bit word arguments.*/
+	Z_LOG_MSG_MODE_SIMPLE,
 };
 
 #define Z_LOG_MSG_DESC_INITIALIZER(_domain_id, _level, _plen, _dlen) \
@@ -226,6 +228,102 @@ enum z_log_msg_mode {
  */
 
 #define Z_LOG_ARM64_VLA_PROTECT() compiler_barrier()
+
+#define _LOG_MSG_SIMPLE_XXXX0 1
+#define _LOG_MSG_SIMPLE_XXXX1 1
+#define _LOG_MSG_SIMPLE_XXXX2 1
+
+/* Determine if amount of arguments (less than 3) qualifies to  simple message. */
+#define LOG_MSG_SIMPLE_ARG_CNT_CHECK(...) \
+	COND_CODE_1(UTIL_CAT(_LOG_MSG_SIMPLE_XXXX, NUM_VA_ARGS_LESS_1(__VA_ARGS__)), (1), (0))
+
+/* Set of marcos used to determine if arguments type allows simplified message creation mode. */
+#define LOG_MSG_SIMPLE_ARG_TYPE_CHECK_0(fmt) 1
+#define LOG_MSG_SIMPLE_ARG_TYPE_CHECK_1(fmt, arg) Z_CBPRINTF_IS_WORD_NUM(arg)
+#define LOG_MSG_SIMPLE_ARG_TYPE_CHECK_2(fmt, arg0, arg1) \
+	Z_CBPRINTF_IS_WORD_NUM(arg0) || Z_CBPRINTF_IS_WORD_NUM(arg1)
+
+/** brief Determine if string arguments types allow to use simplified message creation mode.
+ *
+ * @param ... String with arguments.
+ */
+#define LOG_MSG_SIMPLE_ARG_TYPE_CHECK(...) \
+	UTIL_CAT(LOG_MSG_SIMPLE_ARG_TYPE_CHECK_, NUM_VA_ARGS_LESS_1(__VA_ARGS__))(__VA_ARGS__)
+
+/** @brief Check if message can be handled using simplified method.
+ *
+ * Following conditions must be met:
+ * - 32 bit platform
+ * - Number of arguments from 0 to 2
+ * - Type of an argument must be a numeric value that fits in 32 bit word.
+ *
+ * @param ... String with arguments.
+ *
+ * @retval 1 if message qualifies.
+ * @retval 0 if message does not qualify.
+ */
+#define LOG_MSG_SIMPLE_CHECK(...) \
+	COND_CODE_1(CONFIG_64BIT, (0), (\
+		COND_CODE_1(LOG_MSG_SIMPLE_ARG_CNT_CHECK(__VA_ARGS__), ( \
+				LOG_MSG_SIMPLE_ARG_TYPE_CHECK(__VA_ARGS__)), (0))))
+
+/* Helper macro for handing log with one argument. Macro casts the first argument to uint32_t. */
+#define Z_LOG_MSG_SIMPLE_CREATE_1(_source, _level, ...) \
+	z_log_msg_simple_create_1(_source, _level, GET_ARG_N(1, __VA_ARGS__), \
+			(uint32_t)(uintptr_t)GET_ARG_N(2, __VA_ARGS__))
+
+/* Helper macro for handing log with two arguments. Macro casts arguments to uint32_t.
+ */
+#define Z_LOG_MSG_SIMPLE_CREATE_2(_source, _level, ...) \
+	z_log_msg_simple_create_2(_source, _level, GET_ARG_N(1, __VA_ARGS__), \
+			(uint32_t)(uintptr_t)GET_ARG_N(2, __VA_ARGS__), \
+			(uint32_t)(uintptr_t)GET_ARG_N(3, __VA_ARGS__))
+
+/* Call specific function based on the number of arguments.
+ * Since up 2 to arguments are supported COND_CODE_0 and COND_CODE_1 can be used to
+ * handle all cases (0, 1 and 2 arguments). When tracing is enable then for each
+ * function a macro is create. The difference between function and macro is that
+ * macro is applied to any input arguments so we need to make sure that it is
+ * always called with proper number of arguments. For that it is wrapped around
+ * into another macro and dummy arguments to cover for cases when there is less
+ * arguments in a log call.
+ */
+#define Z_LOG_MSG_SIMPLE_FUNC2(arg_cnt, _source, _level, ...) \
+	COND_CODE_0(arg_cnt, \
+			(z_log_msg_simple_create_0(_source, _level, GET_ARG_N(1, __VA_ARGS__))), \
+			(COND_CODE_1(arg_cnt, ( \
+			    Z_LOG_MSG_SIMPLE_CREATE_1(_source, _level, __VA_ARGS__, dummy) \
+			    ), ( \
+			    Z_LOG_MSG_SIMPLE_CREATE_2(_source, _level, __VA_ARGS__, dummy, dummy) \
+			    ) \
+			)))
+
+/** @brief Call specific function to create a log message.
+ *
+ * Macro picks matching function (based on number of arguments) and calls it.
+ * String arguments are casted to uint32_t.
+ *
+ * @param _source	Source.
+ * @param _level	Severity level.
+ * @param ...		String with arguments.
+ */
+#define LOG_MSG_SIMPLE_FUNC(_source, _level, ...) \
+	Z_LOG_MSG_SIMPLE_FUNC2(NUM_VA_ARGS_LESS_1(__VA_ARGS__), _source, _level, __VA_ARGS__)
+
+/** @brief Create log message using simplified method.
+ *
+ * Macro is gated by the argument count check to run @ref LOG_MSG_SIMPLE_FUNC only
+ * on entries with 2 or less arguments.
+ *
+ * @param _domain_id	Domain ID.
+ * @param _source	Pointer to the source structure.
+ * @param _level	Severity level.
+ * @param ...		String with arguments.
+ */
+#define Z_LOG_MSG_SIMPLE_ARGS_CREATE(_domain_id, _source, _level, ...) \
+	IF_ENABLED(LOG_MSG_SIMPLE_ARG_CNT_CHECK(__VA_ARGS__), (\
+		LOG_MSG_SIMPLE_FUNC(_source, _level, __VA_ARGS__); \
+	))
 
 #define Z_LOG_MSG_STACK_CREATE(_cstr_cnt, _domain_id, _source, _level, _data, _dlen, ...) \
 do { \
@@ -430,6 +528,19 @@ do { \
 					_level, Z_LOG_FMT_ARGS(_fmt, ##__VA_ARGS__)); \
 		_mode = Z_LOG_MSG_MODE_ZERO_COPY; \
 	} else { \
+		IF_ENABLED(UTIL_AND(IS_ENABLED(CONFIG_LOG_SIMPLE_MSG_OPTIMIZE), \
+				    UTIL_AND(UTIL_NOT(_domain_id), UTIL_NOT(_cstr_cnt))), \
+			( \
+			bool can_simple = LOG_MSG_SIMPLE_CHECK(__VA_ARGS__); \
+			if (can_simple && ((_dlen) == 0)) { \
+				LOG_MSG_DBG("create fast message\n");\
+				Z_LOG_MSG_SIMPLE_ARGS_CREATE(_domain_id, _source, _level, \
+						     Z_LOG_FMT_ARGS(_fmt, ##__VA_ARGS__)); \
+				_mode = Z_LOG_MSG_MODE_SIMPLE; \
+				break; \
+			} \
+			) \
+		) \
 		LOG_MSG_DBG("create on stack message\n");\
 		Z_LOG_MSG_STACK_CREATE(_cstr_cnt, _domain_id, _source, _level, _data, \
 					_dlen, Z_LOG_FMT_ARGS(_fmt, ##__VA_ARGS__)); \