--- a/subsys/logging/Kconfig.misc
+++ b/subsys/logging/Kconfig.misc
@@ -33,6 +33,16 @@ config LOG_USE_VLA
 	  supported. Note that VLA are used for arrays which size is resolved at
 	  compile time so at runtime arrays have fixed size.
 
+config LOG_SIMPLE_MSG_OPTIMIZE
+	bool "Optimize simple messages for size"
+	depends on !64BIT && !CBPRINTF_PACKAGE_HEADER_STORE_CREATION_FLAGS
+	default y
+	help
+	  Dedicated code for handling simple log messages (0-2 32 bit word arguments).
+	  Approximately, 70%-80% log messages in the application fit into that category.
+	  Depending on the architecture code size reduction is from 0-40% (highest seen on
+	  RISCV32) and execution time also up to 40%.
+
 config LOG_ALWAYS_RUNTIME
 	bool "Always use runtime message creation (v2)"
 	default y if NO_OPTIMIZATIONS