--- a/subsys/logging/Kconfig.mode
+++ b/subsys/logging/Kconfig.mode
@@ -53,6 +53,12 @@ config LOG_FRONTEND_ONLY
 	  Option indicates that there are no backends intended to be used.
 	  Code asserts if any backend is enabled.
 
+config LOG_FRONTEND_OPT_API
+	bool "Frontend optimized API"
+	help
+	  When enabled, frontend implements functions which are optimized versions
+	  used for the most common log messages.
+
 config LOG_CUSTOM_HEADER
 	bool "Include Custom Log Header"
 	help