It's been noticed that logging messages takes more ROM on RISCV platform others (like ARM Cortex-M). On RISCV32 message takes 70-80 bytes and on Cortex-M 30-40 bytes. This PR is an attempt to improve that by focusing on the most common log messages and optimizing them and using legacy generic approach whenever optimized approach cannot be applied.

I've analyzed zephyr code base (and other projects) which log messages are the most common and got following results:
- 45% of all logs are strings without arguments (e.g. `LOG_INF("foo")`)
- 24% of all logs are strings with one 32 bit argument (e.g.` LOG_INF("char: %c", c)`)
- 6 % of all logs are strings with two 32 bit arguments (e.g. `LOG_INF("%08x %d, a, b)`)

Overall there is usually between 75%-80% of those "simple" messages in the application. On 32 bit platforms messages like that can be built very easily since there is no alignment or padding to consider. Based on that analysis, targeted approach was taken to provide optimized macros and functions for logging "simple" messages on 32 bit architectures.

Optimization is enabled by default (`CONFIG_LOG_SIMPLE_MSG_OPTIMIZE=n` to disable).

Optimization results on RISCV32 (~40% less code)

| Log type    | Before  | After |
| -------------- | ---------- | ---- 
| Log 0 args | 46     | 30 |
| Log 1 arg   | 68     | 38 |
| Log 2 args | 70     | 40 |

Optimization results on ARM Cortex-M3 (~10% less code)

| Log type    | Before  | After |
| -------------- | ---------- | ---- 
| Log 0 args | 36     | 28 |
| Log 1 arg   | 44     | 32 |
| Log 2 args | 44     | 40 |

If we consider RISCV32 application with 300 log messages with average distribution of message types that means 5kB code saving. Improvements on ARM Cortex-M3 are not that significant but they are seen in performance where "simple" logs logging went from **14us** to **10us**. On RISCV32 performance gain is approx.40%.

PR extends also log_frontend API with optional functions for handling "simple" messages.